refactor(tui)!: move all drawing into bitrouter-tui
`apps/bitrouter` no longer depends on `ratatui`. It cannot construct a
widget, so every drawn thing goes through the renderer crate and the
compiler enforces it:

  apps/bitrouter may touch the terminal. Only bitrouter-tui may draw on it.

`crossterm` stays in the app, and that is the rule rather than an
exception to it: crossterm is terminal I/O (owning stdin, bracketed
paste, reading key events), which is this process's; ratatui is drawing,
which is the crate's.

The boundary is drawn on **knowledge, not medium**. Rendering something
is never by itself a reason to keep it in the app; naming something the
protocol does not carry always is.

What moved into the crate

- `picker` — `ProviderInfo` is an `agent-client-protocol-schema` type
  (v1 and v2), and the crate already resolved `unstable_llm_providers`,
  so it compiled there unchanged. This reverts TUI_RENDERER_SPEC §8's
  classification, whose argument — "providers/* is BitRouter's, no
  generic agent serves it" — was about who implements the *server*
  side, which is not what the charter asks. §8 carries an amendment.
- `cost` — split, not moved whole. `Scope{Session,Wider}` + `Cost::new`
  + `render` are protocol-shaped and live in the crate; the vendor
  `_meta` key, its wire spelling, and `from_usage` stay in the app.
  §8 said the split does not compile; that held against the split rev 2
  proposed, not this one.
- `lifecycle` — terminal enter/restore/panic-hook. Custody was split
  across the crate boundary with an ordering contract documented on
  both sides (`Writer::new` performs a DSR query). Signals stay in the
  app: which exist is a property of this process.
- `editor` — the line editor and `is_cancel`/`is_redraw`. What keys
  *mean* is a terminal fact; owning stdin is a process fact, so the
  tokio pump stays in `chat/input.rs`. That split is why the crate still
  needs no async runtime.
- `view` + `plain` — the live view, its footer, and the piped form.
  `View::open` takes a `CostReader`, so the caller must answer "whose
  spend is this?" before a figure can be built. Same shape as
  `Picker::open`'s capability flag: a parameter, never an inference.

Also: `acp_cli.rs` 2597 -> 1767 lines, with the interactive half in
`chat/session.rs`. The cut is at the view-open seam, not the file's
divider, so the launch preamble keeps `Config`, `RouteControl`, and
`resolve_socket_path` while the loop names none of them.

`tui/snapshot.rs`'s `Scope` is deleted: with `as_wire`'s only caller
moved, it had no reader, and `Snapshot.scope` never had one. Note that
`status` therefore still prints an unlabelled spend figure — the
honesty rule `chat` keeps has no equivalent there yet.

BREAKING CHANGE: `bitrouter status --watch` (`-w`) is replaced by
`bitrouter status --requests` (`-r`). The self-refreshing ratatui view is
gone; the new flag prints the same table `--watch` already printed when
piped, identically whether stdout is a terminal or not. Scripts that
piped `--watch` need only the new flag name. The two mutating keys it
offered ran commands you can still run directly: `bitrouter reload`, and
your editor on `bitrouter.yaml`. Use `watch -n1 bitrouter status
--requests` to repeat it.

The live view could not move to `bitrouter-tui`: its rows come from the
metering store and cover every caller, most of which never speak ACP, so
importing them would have put a daemon-wide model inside a
session-scoped crate. ACP has no expression for them either — no model,
latency, or per-request status in v1 or v2, and `Usage`/`Cost` are
cumulative per session. Removing it is what let `ratatui` leave the app.
