+++ apps/bitrouter/Cargo.toml
@@ -54,11 +54,12 @@ agent-client-protocol.workspace = true
 tokio-util = { workspace = true, features = ["compat"] }
 axoupdater = { version = "0.10", default-features = false, features = ["github_releases"] }
 semver = "1"
-# The live view (`bitrouter status --watch`). Not behind a cargo feature: a
-# flag on a documented command that sometimes isn't compiled in costs more
-# confusion than the build time it saves (OBSERVABILITY_TUI_SPEC §6).
+# Everything drawn by this binary. There is deliberately no `ratatui` here:
+# the app may touch the terminal but may not draw on it, and the absence of the
+# dependency is what makes that a compile error rather than a convention.
 bitrouter-tui = { path = "../../crates/bitrouter-tui", version = "=1.0.0-alpha.27" }
-ratatui.workspace = true
+# Terminal *I/O*, which is this process's: owning stdin, bracketed paste, and
+# reading key events for `bitrouter chat`. Not drawing — see above.
 crossterm.workspace = true
 jsonschema = { version = "0.47.0", default-features = false }
 rmcp = { workspace = true, features = ["server", "macros", "transport-io"] }
