+++ apps/bitrouter/src/acp_cli.rs
@@ -665,177 +665,13 @@ pub async fn serve(ctx: SpawnContext<'_>) -> Result<()> {
 
 // ── chat ──────────────────────────────────────────────────────────────────────
 
-/// This process's session log, once the subscriber has opened one.
+/// Launch a session for `agent_id` and hand it to the interactive renderer.
 ///
-/// A global because the path is decided during subscriber init — before any
-/// command is dispatched, and the only moment it exists. A session that ends
-/// badly needs it much later, and threading it through every launch path to
-/// serve the failure case would cost more than it is worth.
-static SESSION_LOG: std::sync::OnceLock<std::path::PathBuf> = std::sync::OnceLock::new();
-
-/// Record where this process's session log lives. Called once at startup.
-pub fn remember_session_log(path: std::path::PathBuf) {
-    let _ = SESSION_LOG.set(path);
-}
-
-/// Show the end of the session log and name the file.
-///
-/// Called only when something went wrong: a permanent pane would cost rows on
-/// every session to serve the rare one that fails. An unreadable log is
-/// reported as unreadable rather than skipped — "there is no log" is the kind
-/// of thing a user needs told, not hidden.
-fn session_log_tail() -> Vec<ratatui::text::Line<'static>> {
-    let Some(path) = SESSION_LOG.get() else {
-        return Vec::new();
-    };
-    let log = match std::fs::read_to_string(path) {
-        Ok(log) => log,
-        Err(e) => format!("(could not read the session log: {e})"),
-    };
-    bitrouter_tui::log_tail::render(path, &log, bitrouter_tui::log_tail::TAIL_LINES)
-}
-
-/// Read the journal, whatever the last holder of the lock did.
-///
-/// A poisoned lock costs a frame at most: the journal is plain data, a panic
-/// mid-`apply` leaves it merely stale, and refusing to draw at all would turn
-/// one lost update into a dead session.
-fn journal(
-    shared: &std::sync::Mutex<bitrouter_tui::journal::Journal>,
-) -> std::sync::MutexGuard<'_, bitrouter_tui::journal::Journal> {
-    match shared.lock() {
-        Ok(journal) => journal,
-        Err(poisoned) => poisoned.into_inner(),
-    }
-}
-
-/// The live view of one chat session.
-///
-/// The document is the journal; this owns the two things the journal cannot
-/// know about — the footer the app composes (§4.5) and the terminal the writer
-/// paints into.
-struct View {
-    writer: bitrouter_tui::writer::Writer<ratatui::backend::CrosstermBackend<std::io::Stdout>>,
-    cache: bitrouter_tui::writer::Cache,
-    registry: bitrouter_tui::render::Registry,
-    /// How this session is routed. Fixed for its life unless `/route` changes
-    /// it, and absent when the session is direct.
-    route: Option<String>,
-    /// The most recent thing the client itself has to say — a stop reason, a
-    /// route change, a failed turn, the agent's command list. Replaced rather
-    /// than accumulated: it is about the last thing that happened, and the
-    /// session log is where the history lives.
-    notice: Vec<ratatui::text::Line<'static>>,
-    /// A row owned by an open modal. Today only the provider picker.
-    modal: Option<ratatui::text::Line<'static>>,
-    /// The line being typed. Raw mode means nothing is echoed unless the
-    /// footer echoes it.
-    input: String,
-}
-
-impl View {
-    /// Take the terminal.
-    ///
-    /// Called **before** the session's stdin owner exists: the writer reads
-    /// the cursor once at construction, which on a real terminal is a DSR
-    /// query, and a reader already sitting on stdin would take the answer.
-    fn open(route: Option<String>) -> Result<Self> {
-        let backend = ratatui::backend::CrosstermBackend::new(std::io::stdout());
-        Ok(Self {
-            writer: bitrouter_tui::writer::Writer::new(backend).context("opening the writer")?,
-            cache: bitrouter_tui::writer::Cache::default(),
-            registry: bitrouter_tui::render::Registry::default(),
-            route,
-            notice: Vec::new(),
-            modal: None,
-            input: String::new(),
-        })
-    }
-
-    /// The rows below the document: what this session costs and where it is
-    /// going, then whatever is currently being asked or typed.
-    ///
-    /// Re-emitted as the document's tail on every frame, so it is always the
-    /// newest content and always in the live region — which is what bounds the
-    /// stale-row cost of §4.3 to individual scrolled-off tool calls rather
-    /// than to the session's own summary.
-    fn footer(
-        &self,
-        journal: &bitrouter_tui::journal::Journal,
-    ) -> Vec<ratatui::text::Line<'static>> {
-        use ratatui::text::{Line, Span};
-
-        let cost = journal
-            .usage()
-            .and_then(crate::chat::cost::Cost::from_usage)
-            .map_or_else(crate::chat::cost::unreported, |cost| cost.render());
-        let mut status: Vec<Span<'static>> = cost.spans;
-        if let Some(route) = &self.route {
-            status.push(Span::raw(format!(" · via {route}")));
-        }
-        // Mode, configuration and title are the session's own state, so they
-        // are rendered from the protocol by the crate that owns the protocol.
-        status.extend(bitrouter_tui::render::session::state(
-            journal.mode(),
-            journal.config(),
-            journal.title(),
-        ));
-
-        let mut rows = vec![Line::from(status)];
-        rows.extend(self.notice.iter().cloned());
-        // A question waiting on a person outranks anything else down here.
-        if let Some(prompt) = journal.pending_permission() {
-            rows.push(prompt.render());
-        }
-        if let Some(modal) = &self.modal {
-            rows.push(modal.clone());
-        }
-        rows.push(Line::from(format!("> {}", self.input)));
-        rows
-    }
-
-    /// Paint one frame.
-    ///
-    /// Every caller of this has already decided the frame is worth painting —
-    /// the writer still writes nothing when the document is unchanged, so a
-    /// redundant call costs a render and no output.
-    fn paint(&mut self, shared: &std::sync::Mutex<bitrouter_tui::journal::Journal>) -> Result<()> {
-        let size = self.writer.size();
-        let journal = journal(shared);
-        let footer = self.footer(&journal);
-        let document = self.cache.document(&journal, &self.registry, size, &footer);
-        // The lock is not held across the write: a frame can take a
-        // millisecond, and the pump has updates to apply.
-        drop(journal);
-        self.writer.frame(&document).context("painting a frame")
-    }
-
-    /// Repaint everything on screen, whatever the writer believes is there.
-    ///
-    /// For when something else has written to this terminal: the writer paints
-    /// against its own model and never asks the terminal what it holds, so a
-    /// stray line leaves every row below it one off. `Ctrl-L` is how a person
-    /// says so.
-    fn redraw(&mut self, shared: &std::sync::Mutex<bitrouter_tui::journal::Journal>) -> Result<()> {
-        self.writer.invalidate();
-        self.paint(shared)
-    }
-
-    /// Leave the cursor below the document. Nothing rendered is removed.
-    fn finish(&mut self) -> Result<()> {
-        self.writer.finish().context("closing the view")
-    }
-}
-
-/// Run one ACP session interactively in the caller's terminal.
-///
-/// The interactive counterpart to [`serve`]: same launch, same routing, same
-/// per-session log — but the session is rendered for a person instead of
-/// exposed to a manager over stdio.
-///
-/// This owns argument handling, routing, and session lifetime; the rendering
-/// belongs to `bitrouter-tui`, which cannot depend on this crate and so is
-/// handed only what ACP carries.
+/// This half is the composition root: it resolves routing, launches the agent,
+/// attaches observability, and builds the routing surface — all of which need
+/// `Config`, the config source, and the daemon's control socket. What it hands
+/// over needs none of them, which is why the loop itself lives in
+/// [`crate::chat::session`].
 pub async fn chat(ctx: SpawnContext<'_>) -> Result<()> {
     use std::io::IsTerminal as _;
 
@@ -869,638 +705,40 @@ pub async fn chat(ctx: SpawnContext<'_>) -> Result<()> {
     // those escapes into a file or a `| grep` would corrupt the very output
     // the redirect exists to capture. So a redirected stdout gets the session
     // as plain text instead, which is the one thing a pipe *can* use.
-    if !std::io::stdout().is_terminal() {
-        let ended = chat_plain(session).await;
-        if let Some(exporter) = exporter {
-            exporter.shutdown();
-        }
-        return ended;
-    }
-
-    // The picker exists only when this session can actually be rerouted, which
-    // needs both a daemon and an attributable launch (task 4.2). Probing that
-    // here rather than assuming it is what keeps a dead control off the screen.
-    let providers = SessionProviders::new(
-        &config,
-        routed
-            .via
-            .as_ref()
-            .and(routed.launch_id.as_ref())
-            .map(|id| RouteControl {
-                socket: crate::daemon::resolve_socket_path(
-                    source.home().join("bitrouter.yaml").as_path(),
-                    &config.server.control_socket,
-                ),
-                launch_id: id.clone(),
-            }),
-    );
-    let routable = providers.can_reroute();
-    if routable {
-        eprintln!("chat: type /route to change provider mid-session.");
-    }
-    eprintln!("chat: type a message and press enter; Ctrl-D to end the session.");
-
-    // The view opens **before** the stdin owner: the writer reads the cursor
-    // once, here, and on a real terminal that is a DSR query whose answer a
-    // reader already sitting on stdin would take.
-    let mut view = View::open(routed.via.clone())?;
-
-    // Raw mode starts here, after the last plain `eprintln!` — a cooked
-    // newline in a raw terminal does not return the carriage. From this point
-    // the terminal echoes nothing and delivers no SIGINT: both are ours.
-    let mut stdin = crate::chat::input::Stdin::open().context("taking the terminal for input")?;
-    // The three exits, all landing at `lifecycle::restore()`: `Stdin`'s drop
-    // covers the normal one and every `?` on the way out, the panic hook
-    // covers the second, and `Shutdown` is the third — a signal runs no Rust
-    // the loop controls, so it has to be awaited rather than caught.
-    crate::tui::lifecycle::install_panic_restore();
-    let mut shutdown = crate::tui::lifecycle::Shutdown::install();
-
-    // Permission requests block the turn until a person answers. They are the
-    // journal's, but they arrive on their own channel rather than as updates.
-    let (permission_tx, mut permission_rx) = tokio::sync::mpsc::unbounded_channel();
-    let mut pending = session.permissions();
-    let permissions = tokio::spawn(async move {
-        while let Some(request) = pending.next().await {
-            if permission_tx.send(request).is_err() {
-                break;
-            }
-        }
-    });
-
-    // The **raw** ACP stream, not the translated one: the journal is a
-    // protocol client, and translating first would lose exactly the fidelity
-    // it exists to retain.
-    let mut updates = session.raw_updates();
-    let shared = std::sync::Arc::new(std::sync::Mutex::new(
-        bitrouter_tui::journal::Journal::default(),
-    ));
-    // A signal, not a payload: the pump has already applied the update by the
-    // time this arrives, so the loop's only job is to decide when to paint.
-    let (dirty_tx, mut dirty_rx) = tokio::sync::mpsc::unbounded_channel();
-    let pump_journal = std::sync::Arc::clone(&shared);
-    let pump = tokio::spawn(async move {
-        while let Some(update) = updates.next().await {
-            journal(&pump_journal).apply(update);
-            if dirty_tx.send(()).is_err() {
-                return;
-            }
-        }
-    });
-
-    // Every prompt gets its own id so two in a row cannot merge into one run —
-    // which they otherwise would, if the agent answered the first with
-    // nothing at all.
-    let mut prompts = 0_usize;
-    let mut abnormal = false;
-
-    'session: loop {
-        view.input.clear();
-        view.paint(&shared)?;
-        let read = tokio::select! {
-            // Raw mode means nothing is echoed unless the footer echoes it.
-            read = stdin.read_line(|echo| match echo {
-                crate::chat::input::Echo::Changed(typed) => {
-                    view.input = typed.to_string();
-                    view.paint(&shared)
-                }
-                crate::chat::input::Echo::Redraw(typed) => {
-                    view.input = typed.to_string();
-                    view.redraw(&shared)
-                }
-            }) => read?,
-            // A signal at an idle prompt ends the session the same way Ctrl-D
-            // does — through teardown, not around it.
-            () = shutdown.recv() => break 'session,
-        };
-        let line = match read {
-            crate::chat::input::Prompt::Line(line) => line,
-            crate::chat::input::Prompt::End => break,
-        };
-        view.input.clear();
-        // The last turn's word stands until this one starts, so a stop reason
-        // is readable for as long as the reader is deciding what to say next
-        // rather than for the one frame between them.
-        view.notice.clear();
-        // A line of exactly `/commands` lists what the agent itself offers.
-        // Ours are hardcoded; the agent's arrive on `AvailableCommandsUpdate`
-        // and were invisible until now.
-        if line.trim() == "/commands" {
-            view.notice = bitrouter_tui::render::session::commands(journal(&shared).commands());
-            view.paint(&shared)?;
-            continue;
-        }
-        // A line of exactly `/route` opens the picker. Only offered when the
-        // session can honour it — see `can_reroute`.
-        if line.trim() == "/route" {
-            if routable {
-                pick_provider(&mut view, &shared, &mut stdin, &providers).await?;
-            } else {
-                view.notice = vec![ratatui::text::Line::from(
-                    "this session cannot be rerouted (running direct, or its credential is \
-                     its own and cannot be attributed)",
-                )];
-                view.paint(&shared)?;
-            }
-            continue;
-        }
-        // The prompt joins the document in the user's own voice, so it scrolls
-        // with the answer it asked for.
-        prompts = prompts.saturating_add(1);
-        journal(&shared).apply(SessionUpdate::UserMessageChunk(prompt_chunk(
-            &line, prompts,
-        )));
-        view.paint(&shared)?;
-
-        let turn = session.prompt(&line);
-        tokio::pin!(turn);
-        // The tick is the streaming frame budget. It lives here rather than at
-        // the prompt because streaming is the only thing that needs
-        // coalescing: a keystroke and a permission paint immediately.
-        let mut ticker = tokio::time::interval(bitrouter_tui::writer::Schedule::INTERVAL);
-        ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-        let mut schedule = bitrouter_tui::writer::Schedule::default();
-        // Both are handled *after* the `select!` rather than inside a branch:
-        // the branch bodies run while the arms' futures are still borrowed,
-        // and one of those arms holds `stdin` — which is exactly what
-        // answering a permission needs.
-        let mut requested = None;
-        let mut cancelled = false;
-        let mut redraw = false;
-        let outcome = loop {
-            let mut paint = false;
-            tokio::select! {
-                update = dirty_rx.recv() => match update {
-                    // Already in the journal; the tick decides when it is seen.
-                    Some(()) => {
-                        schedule.wake(
-                            bitrouter_tui::writer::Trigger::Update,
-                            std::time::Instant::now(),
-                        );
-                    }
-                    None => continue,
-                },
-                _ = ticker.tick() => {
-                    paint = schedule.wake(
-                        bitrouter_tui::writer::Trigger::Tick,
-                        std::time::Instant::now(),
-                    );
-                },
-                request = permission_rx.recv() => match request {
-                    Some(request) => requested = Some(request),
-                    None => continue,
-                },
-                event = stdin.next_event() => match event {
-                    // Ctrl-C and `Esc` during a turn are a cancel, not an
-                    // exit: the session survives it and the next prompt is
-                    // drawn. No modal can be open here — one would own the key
-                    // stream while it ran — so `Esc` has nothing else to close.
-                    Some(event) => {
-                        cancelled = crate::chat::input::is_cancel(&event);
-                        redraw = crate::chat::input::is_redraw(&event);
-                    }
-                    // The terminal went away. Nothing can answer the rest of
-                    // this turn, so stop waiting on it.
-                    None => cancelled = true,
-                },
-                result = &mut turn => break Some(result),
-                // Mid-turn, a signal still leaves by the front door: the agent
-                // is shut down and the terminal restored on the way out.
-                () = shutdown.recv() => break 'session,
-            }
-            if std::mem::take(&mut redraw) {
-                view.redraw(&shared)?;
-                schedule.wake(
-                    bitrouter_tui::writer::Trigger::Key,
-                    std::time::Instant::now(),
-                );
-            } else if paint {
-                view.paint(&shared)?;
-            }
-            if let Some(request) = requested.take() {
-                if cancelled {
-                    // Cancelling with a question outstanding answers it — the
-                    // one thing it must never do is leave it to be answered by
-                    // a keystroke meant for something else.
-                    deny(request);
-                } else {
-                    answer_permission(&mut view, &shared, &mut stdin, request).await?;
-                    schedule.wake(
-                        bitrouter_tui::writer::Trigger::Permission,
-                        std::time::Instant::now(),
-                    );
-                }
-            }
-            if cancelled {
-                break None;
-            }
-        };
-        if outcome.is_none() {
-            // Tell the agent, not merely ourselves: dropping the turn future
-            // stops this side waiting and leaves the agent working.
-            let told = session.cancel().await;
-            // And leave no question hanging. A permission the agent asked
-            // after the cancel has nobody to answer it, and an unanswered
-            // question must resolve to deny — never to consent, and never by
-            // sitting there until a later keystroke picks an option.
-            while let Ok(request) = permission_rx.try_recv() {
-                deny(request);
-            }
-            journal(&shared).set_pending_permission(None);
-            if let Err(e) = told {
-                tracing::warn!(error = %e, "cancelling the turn");
-            }
-        }
-        // Whatever the pump applied between its last signal and the turn
-        // resolving is already in the journal; draining the signals is what
-        // makes the frame below include it.
-        while dirty_rx.try_recv().is_ok() {}
-        view.notice = vec![match outcome {
-            None => ratatui::text::Line::from("[turn cancelled]"),
-            Some(Ok(response)) => {
-                ratatui::text::Line::from(format!("[{:?}]", response.stop_reason))
-            }
-            Some(Err(e)) => {
-                // A failed turn is the abnormal exit the log tail exists for;
-                // it is written after teardown, when the terminal is the
-                // shell's again.
-                abnormal = true;
-                ratatui::text::Line::from(format!("turn failed: {e}"))
-            }
-        }];
-        // A settled turn is immediate: it is the moment the reader is waiting
-        // for, not streaming noise.
-        schedule.wake(
-            bitrouter_tui::writer::Trigger::TurnSettled,
-            std::time::Instant::now(),
+    let ended = if std::io::stdout().is_terminal() {
+        // The picker exists only when this session can actually be rerouted,
+        // which needs both a daemon and an attributable launch (task 4.2).
+        // Probing that here rather than assuming it is what keeps a dead
+        // control off the screen — and it is probed here rather than in the
+        // renderer because the daemon socket is this half's to know.
+        let providers = SessionProviders::new(
+            &config,
+            routed
+                .via
+                .as_ref()
+                .and(routed.launch_id.as_ref())
+                .map(|id| RouteControl {
+                    socket: crate::daemon::resolve_socket_path(
+                        source.home().join("bitrouter.yaml").as_path(),
+                        &config.server.control_socket,
+                    ),
+                    launch_id: id.clone(),
+                }),
         );
-        view.paint(&shared)?;
-    }
+        let routable = providers.can_reroute();
+        if routable {
+            eprintln!("chat: type /route to change provider mid-session.");
+        }
+        eprintln!("chat: type a message and press enter; Ctrl-D to end the session.");
+        crate::chat::session::run(session, providers, routable, routed.via.clone()).await
+    } else {
+        crate::chat::session::chat_plain(session).await
+    };
 
-    pump.abort();
-    permissions.abort();
-    // A session whose agent could not be shut down cleanly ended abnormally
-    // too, and the log is where the reason is.
-    let shutdown = session.shutdown().await;
-    abnormal = abnormal || shutdown.is_err();
-    // Leave the cursor below the document, then give the terminal back —
-    // in that order, because the log tail below is written as ordinary
-    // output and must land in a terminal that is the shell's again.
-    view.finish()?;
-    drop(stdin);
-    if abnormal {
-        write_plain(&mut std::io::stdout(), &session_log_tail())?;
-    }
-    shutdown.context("shutting down chat session")?;
     if let Some(exporter) = exporter {
         exporter.shutdown();
     }
-    Ok(())
-}
-
-/// Answer a permission nobody is going to answer.
-///
-/// A turn can be cancelled with a question outstanding, and a cancelled turn
-/// must never resolve to consent. This is the same `Prompt::deny()` path an
-/// `Esc` at the prompt takes, so there is exactly one rule for "no answer" and
-/// it is the safe one.
-fn deny(request: bitrouter_sdk::acp::up::PendingPermission) {
-    let prompt = bitrouter_tui::permission::Prompt::new(
-        request.tool_call.fields.title.clone(),
-        request.tool_call.tool_call_id.0.to_string(),
-        request.options.clone(),
-    );
-    request.resolve(unanswered(&prompt));
-}
-
-/// What an unanswered permission resolves to.
-///
-/// An explicit reject when the agent offered one, so it hears a decision it
-/// understands. Otherwise **cancelled** — never a selection, because the only
-/// options left would be ones that say yes.
-fn unanswered(
-    prompt: &bitrouter_tui::permission::Prompt,
-) -> agent_client_protocol::schema::v1::RequestPermissionOutcome {
-    use agent_client_protocol::schema::v1::{RequestPermissionOutcome, SelectedPermissionOutcome};
-
-    match prompt.deny() {
-        Some(id) => RequestPermissionOutcome::Selected(SelectedPermissionOutcome::new(id)),
-        None => RequestPermissionOutcome::Cancelled,
-    }
-}
-
-/// The user's own prompt, as the update the agent would have sent for it.
-///
-/// Keyed, because the journal's sticky rule continues an open run on an
-/// unkeyed chunk: two prompts in a row with no answer between them would
-/// otherwise become one paragraph.
-fn prompt_chunk(line: &str, nth: usize) -> agent_client_protocol::schema::v1::ContentChunk {
-    use agent_client_protocol::schema::v1::{ContentBlock, ContentChunk, MessageId, TextContent};
-
-    let mut chunk = ContentChunk::new(ContentBlock::Text(TextContent::new(line.to_string())));
-    chunk.message_id = Some(MessageId::from(format!("chat:prompt:{nth}")));
-    chunk
-}
-
-/// A rendered line as plain text: the spans concatenated, every style
-/// dropped.
-///
-/// The styles are the whole difference between the two paths. A `Line` is
-/// only escape sequences once a backend writes it; written this way it never
-/// becomes any.
-fn plain_text(line: &ratatui::text::Line<'static>) -> String {
-    line.spans
-        .iter()
-        .map(|span| span.content.as_ref())
-        .collect()
-}
-
-fn write_plain(
-    out: &mut impl std::io::Write,
-    lines: &[ratatui::text::Line<'static>],
-) -> Result<()> {
-    for line in lines {
-        // `writeln!` rather than `println!`: a closed pipe makes `println!`
-        // panic, and `bitrouter chat … | head` closes the pipe as a matter of
-        // course.
-        writeln!(out, "{}", plain_text(line)).context("writing the session to stdout")?;
-    }
-    Ok(())
-}
-
-/// The same session, for a stdout that is not a terminal.
-///
-/// Deliberately not a second renderer: it is the same journal and the same
-/// renderers, printed without a backend. What it drops is everything that
-/// needs a screen — the footer, the picker, raw mode, painting in place — and
-/// what it keeps is the transcript, which is the part a pipe can use.
-///
-/// The document is written **once per turn**, from the row after the last one
-/// written. A pipe cannot take a row back, so nothing is emitted until the
-/// turn that produces it has settled — which is also what makes in-place
-/// patching arrive as one finished tool call rather than three.
-///
-/// A permission request is denied rather than asked, because there is nobody
-/// to ask: the terminal that would carry the question is the one that isn't
-/// there. Denying is the same answer this path gives an unanswerable prompt
-/// anywhere else, and it is never mistaken for consent.
-async fn chat_plain(session: bitrouter_sdk::acp::engine::Session) -> Result<()> {
-    use std::io::Write as _;
-
-    use agent_client_protocol::schema::v1::{RequestPermissionOutcome, SelectedPermissionOutcome};
-    use futures::FutureExt as _;
-    use tokio::io::AsyncBufReadExt as _;
-
-    /// The terminal a pipe does not have. Eighty columns is the convention
-    /// every tool that has had to guess has guessed.
-    const PIPED: ratatui::layout::Size = ratatui::layout::Size {
-        width: 80,
-        height: 24,
-    };
-
-    let mut out = std::io::stdout();
-    let mut journal = bitrouter_tui::journal::Journal::default();
-    let mut cache = bitrouter_tui::writer::Cache::default();
-    let registry = bitrouter_tui::render::Registry::default();
-    // How many rows of the document have already been written.
-    let mut written = 0_usize;
-    let mut prompts = 0_usize;
-    let mut updates = session.raw_updates();
-    let mut permissions = session.permissions();
-    // The only reader of stdin on this path, and it takes no raw mode — the
-    // one that does (`chat::input::Stdin`) needs a terminal, which is exactly
-    // what is missing here.
-    let mut lines = tokio::io::BufReader::new(tokio::io::stdin()).lines();
-
-    let ended = loop {
-        let Some(line) = lines.next_line().await.context("reading stdin")? else {
-            break Ok(());
-        };
-        if line.trim().is_empty() {
-            continue;
-        }
-        if line.trim() == "/route" {
-            writeln!(out, "/route needs a terminal").context("writing to stdout")?;
-            continue;
-        }
-        prompts = prompts.saturating_add(1);
-        journal.apply(SessionUpdate::UserMessageChunk(prompt_chunk(
-            &line, prompts,
-        )));
-
-        let turn = session.prompt(&line);
-        tokio::pin!(turn);
-        let outcome = loop {
-            tokio::select! {
-                update = updates.next() => if let Some(update) = update {
-                    journal.apply(update);
-                },
-                request = permissions.next() => if let Some(request) = request {
-                    let prompt = bitrouter_tui::permission::Prompt::new(
-                        request.tool_call.fields.title.clone(),
-                        request.tool_call.tool_call_id.0.to_string(),
-                        request.options.clone(),
-                    );
-                    request.resolve(match prompt.deny() {
-                        Some(id) => RequestPermissionOutcome::Selected(
-                            SelectedPermissionOutcome::new(id),
-                        ),
-                        None => RequestPermissionOutcome::Cancelled,
-                    });
-                    writeln!(out, "  permission denied: no terminal to ask on")
-                        .context("writing the permission decision")?;
-                },
-                result = &mut turn => break result,
-            }
-        };
-        // Whatever the agent emitted between its last update and the turn
-        // resolving. There is nothing to flush: the journal has no buffered
-        // state, which is the whole reason it replaced `Transcript`.
-        while let Some(Some(update)) = updates.next().now_or_never() {
-            journal.apply(update);
-        }
-        let document = cache.document(&journal, &registry, PIPED, &[]);
-        write_plain(&mut out, document.get(written..).unwrap_or_default())?;
-        written = document.len();
-        match outcome {
-            Ok(response) => {
-                writeln!(out, "[{:?}]", response.stop_reason).context("writing the stop reason")?
-            }
-            Err(e) => {
-                writeln!(out, "turn failed: {e}").context("writing the failure")?;
-                write_plain(&mut out, &session_log_tail())?;
-            }
-        }
-    };
-
-    let shutdown = session.shutdown().await;
-    if shutdown.is_err() {
-        write_plain(&mut out, &session_log_tail())?;
-    }
-    ended.and(shutdown.context("shutting down chat session"))
-}
-
-/// Draw the provider picker and, on a selection, actually change the route.
-///
-/// The order matters: `providers/set` is issued first and the list is re-read
-/// afterwards, so what the user ends up seeing is the route the daemon is
-/// serving rather than the one they asked for. A `set` that fails leaves the
-/// old route marked, and says why.
-async fn pick_provider(
-    view: &mut View,
-    shared: &std::sync::Mutex<bitrouter_tui::journal::Journal>,
-    stdin: &mut crate::chat::input::Stdin,
-    providers: &SessionProviders,
-) -> Result<()> {
-    use bitrouter_sdk::acp::down::ProviderSurface;
-    use crossterm::event::{Event, KeyCode, KeyEventKind, KeyModifiers};
-
-    let listed = providers.list().await;
-    let Some(picker) = crate::chat::picker::Picker::open(true, &listed) else {
-        view.notice = vec![ratatui::text::Line::from(
-            "no routable providers to choose between",
-        )];
-        view.paint(shared)?;
-        return Ok(());
-    };
-    // The picker is a footer row for as long as it is open, repainted in
-    // place like everything else down there.
-    view.modal = Some(picker.render());
-    view.paint(shared)?;
-
-    // Keys come from the session's one stdin owner, which already holds raw
-    // mode: a modal that took and dropped it would be a second owner, and the
-    // window between them is where a keystroke gets lost.
-    let chosen = loop {
-        match stdin.next_event().await {
-            Some(Event::Key(key)) if key.kind == KeyEventKind::Press => {
-                // A control chord is never a choice — Ctrl-C closes the
-                // picker instead of selecting whatever `c` happens to be.
-                if key.modifiers.contains(KeyModifiers::CONTROL) {
-                    break None;
-                }
-                match key.code {
-                    KeyCode::Char(c) => {
-                        if let Some(id) = picker.choose(c) {
-                            break Some(id);
-                        }
-                    }
-                    KeyCode::Esc => break None,
-                    _ => {}
-                }
-            }
-            // The terminal went away mid-question; the route is unchanged.
-            None => break None,
-            Some(_) => {}
-        }
-    };
-
-    view.modal = None;
-    let Some(id) = chosen else {
-        view.notice = vec![ratatui::text::Line::from("route unchanged")];
-        view.paint(shared)?;
-        return Ok(());
-    };
-
-    // Attempt it, then report what is actually in force — never what was asked
-    // for. `providers/set` can legitimately refuse.
-    let request = agent_client_protocol::schema::v1::SetProviderRequest::new(
-        agent_client_protocol::schema::v1::ProviderId::new(id.clone()),
-        LlmProtocol::Other(String::new()),
-        String::new(),
-    );
-    match providers.set(request).await {
-        Ok(()) => {
-            let confirmed = providers.list().await;
-            let in_force = confirmed
-                .iter()
-                .find(|p| p.current.is_some())
-                .map(|p| p.provider_id.0.to_string())
-                .unwrap_or_else(|| "unchanged".to_string());
-            view.notice = vec![ratatui::text::Line::from(format!("route: {in_force}"))];
-            // The footer names the route for the rest of the session, not just
-            // for this frame.
-            view.route = Some(in_force);
-        }
-        Err(e) => {
-            view.notice = vec![ratatui::text::Line::from(format!("route unchanged: {e}"))];
-        }
-    }
-    view.paint(shared)
-}
-
-/// Draw a permission request and block on the user's answer.
-///
-/// Reads keys directly rather than through the line-based stdin loop: a
-/// permission answer is one keystroke, and requiring enter would make the
-/// fast path — deny — slower than the dangerous one.
-///
-/// Every path that is not an explicit choice resolves to deny or cancel.
-/// A prompt that resolved ambiguity as consent would be worse than no prompt.
-async fn answer_permission(
-    view: &mut View,
-    shared: &std::sync::Mutex<bitrouter_tui::journal::Journal>,
-    stdin: &mut crate::chat::input::Stdin,
-    request: bitrouter_sdk::acp::up::PendingPermission,
-) -> Result<()> {
-    use agent_client_protocol::schema::v1::{RequestPermissionOutcome, SelectedPermissionOutcome};
-    use crossterm::event::{Event, KeyCode, KeyEventKind, KeyModifiers};
-
-    let prompt = bitrouter_tui::permission::Prompt::new(
-        request.tool_call.fields.title.clone(),
-        request.tool_call.tool_call_id.0.to_string(),
-        request.options.clone(),
-    );
-    // The journal holds the open question, so every frame drawn while it is
-    // open shows it — including one painted by something else entirely.
-    journal(shared).set_pending_permission(Some(prompt.clone()));
-    view.paint(shared)?;
-
-    let deny = || unanswered(&prompt);
-
-    // Keys come from the session's one stdin owner. It already holds raw mode
-    // for the whole session, so there is no mode to take here and none to
-    // leave behind if this future is dropped.
-    let outcome = loop {
-        match stdin.next_event().await {
-            Some(Event::Key(key)) if key.kind == KeyEventKind::Press => {
-                // Ctrl-C answers the question the only way an interrupt can be
-                // read: no. Passing the chord's letter to `choose` could
-                // select an option, which is the one outcome a cancel must
-                // never produce.
-                if key.modifiers.contains(KeyModifiers::CONTROL) {
-                    break deny();
-                }
-                match key.code {
-                    KeyCode::Char(c) => {
-                        if let Some(id) = prompt.choose(c) {
-                            break RequestPermissionOutcome::Selected(
-                                SelectedPermissionOutcome::new(id),
-                            );
-                        }
-                    }
-                    KeyCode::Esc => break deny(),
-                    _ => {}
-                }
-            }
-            // The terminal went away mid-question. Deny — an unanswerable
-            // prompt must not become an allow.
-            None => break deny(),
-            Some(_) => {}
-        }
-    };
-
-    let chosen = matches!(outcome, RequestPermissionOutcome::Selected(_));
-    request.resolve(outcome);
-    // Answered: the question stops being asked, and what was decided is said
-    // once rather than left on screen as though it were still open.
-    journal(shared).set_pending_permission(None);
-    view.notice = vec![ratatui::text::Line::from(if chosen {
-        "permission answered"
-    } else {
-        "permission denied"
-    })];
-    view.paint(shared)
+    ended
 }
 
 // ── prompt ────────────────────────────────────────────────────────────────────
@@ -2101,7 +1339,7 @@ async fn measured_usage_update(
     record: &RequestCompleted,
 ) -> Option<SessionUpdate> {
     use crate::metering::store::TimeWindow;
-    use crate::tui::snapshot::Scope;
+    use bitrouter_tui::cost::Scope;
 
     let window = TimeWindow::Custom {
         start: session_start,
@@ -2114,9 +1352,9 @@ async fn measured_usage_update(
     let (summary, scope) = match launch_id {
         Some(id) => (
             store.spend_summary_for_launch(id, window).await.ok()?,
-            Scope::Launch,
+            Scope::Session,
         ),
-        None => (store.spend_summary(window).await.ok()?, Scope::DaemonWide),
+        None => (store.spend_summary(window).await.ok()?, Scope::Wider),
     };
     if summary.requests == 0 {
         return None;
@@ -2129,7 +1367,7 @@ async fn measured_usage_update(
     let mut meta = serde_json::Map::new();
     meta.insert(
         crate::chat::cost::COST_SCOPE_META_KEY.to_string(),
-        serde_json::Value::String(scope.as_wire().to_string()),
+        serde_json::Value::String(crate::chat::cost::to_wire(scope).to_string()),
     );
     Some(SessionUpdate::UsageUpdate(
         UsageUpdate::new(
@@ -2168,74 +1406,6 @@ pub fn launch_options(turn_timeout_secs: Option<u64>) -> LaunchOptions {
 pub(crate) fn catalog_from_config(config: &Config) -> Result<ConfigAcpRoutingTable> {
     ConfigAcpRoutingTable::from_configs(config.agents.iter().map(|(k, v)| (k.clone(), v.clone())))
         .context("building acp routing table from config")
-}
-
-#[cfg(test)]
-mod cancel_tests {
-    use agent_client_protocol::schema::v1::{
-        PermissionOption, PermissionOptionId, PermissionOptionKind, RequestPermissionOutcome,
-    };
-
-    use super::*;
-
-    fn prompt(options: Vec<PermissionOption>) -> bitrouter_tui::permission::Prompt {
-        bitrouter_tui::permission::Prompt::new(Some("Write src/main.rs".to_string()), "t1", options)
-    }
-
-    fn option(id: &str, kind: PermissionOptionKind) -> PermissionOption {
-        PermissionOption::new(
-            PermissionOptionId::new(id.to_string()),
-            id.to_string(),
-            kind,
-        )
-    }
-
-    /// The rule a cancelled turn depends on: a question nobody answered
-    /// resolves to the agent's own reject option.
-    ///
-    /// Cancelling is not consenting. A turn can be cancelled with a permission
-    /// outstanding — `Esc` or Ctrl-C while the agent is asking — and the
-    /// cancel path answers it here rather than leaving it for whichever key
-    /// happens to arrive next.
-    #[test]
-    fn an_unanswered_permission_takes_the_reject_option() {
-        let offered = prompt(vec![
-            option("allow", PermissionOptionKind::AllowOnce),
-            option("allow-always", PermissionOptionKind::AllowAlways),
-            option("reject", PermissionOptionKind::RejectOnce),
-        ]);
-        let chosen = match unanswered(&offered) {
-            RequestPermissionOutcome::Selected(selected) => Some(selected.option_id.0.to_string()),
-            // `Cancelled`, or a variant added after this build — either way,
-            // not a selection.
-            _ => None,
-        };
-        assert_eq!(
-            chosen.as_deref(),
-            Some("reject"),
-            "the reject option, not the first one offered"
-        );
-    }
-
-    /// And when the agent offered no way to say no, the answer is **cancelled**
-    /// — never one of the options, because every option left says yes.
-    #[test]
-    fn an_unanswered_permission_never_resolves_to_consent() {
-        let only_yes = prompt(vec![
-            option("allow", PermissionOptionKind::AllowOnce),
-            option("allow-always", PermissionOptionKind::AllowAlways),
-        ]);
-        assert!(
-            matches!(unanswered(&only_yes), RequestPermissionOutcome::Cancelled),
-            "an unanswerable question must not become an allow"
-        );
-
-        // Nor when the agent offered nothing at all.
-        assert!(matches!(
-            unanswered(&prompt(Vec::new())),
-            RequestPermissionOutcome::Cancelled
-        ));
-    }
 }
 
 #[cfg(test)]
